# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "create a new customer",
                "parameters": [
                    {
                        "description": "customer data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/get/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "get customer by id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.CustomerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/getAll": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "get all customers details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.CustomerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/remove/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "delete a customer from the system",
                "parameters": [
                    {
                        "type": "string",
                        "description": "customer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/update": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "update an existing customer",
                "parameters": [
                    {
                        "description": "customer data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "isSuccess": {
                    "type": "boolean",
                    "default": true
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.CreateCustomerRequest": {
            "type": "object",
            "required": [
                "code",
                "companyName"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "models.CustomerResponse": {
            "type": "object",
            "properties": {
                "EDI": {
                    "type": "string"
                },
                "GLN": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "bankAccountNumber": {
                    "type": "string"
                },
                "bankIBAN": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "bankSWIFT": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "colorStatus": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "companyTypeID": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "countryID": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "creditCardLastNumbers": {
                    "type": "string"
                },
                "customerBalanceDisabled": {
                    "type": "integer"
                },
                "customerCardNumber": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "customerType": {
                    "type": "string"
                },
                "defaultAssociationID": {
                    "type": "integer"
                },
                "defaultAssociationName": {
                    "type": "string"
                },
                "defaultProfessionalID": {
                    "type": "integer"
                },
                "defaultProfessionalName": {
                    "type": "string"
                },
                "eInvoiceEmail": {
                    "type": "string"
                },
                "eInvoiceEnabled": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "emailEnabled": {
                    "type": "integer"
                },
                "emailOptOut": {
                    "type": "integer"
                },
                "euCustomerType": {
                    "type": "string"
                },
                "facebookName": {
                    "type": "string"
                },
                "factoringContractNumber": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "flagStatus": {
                    "type": "integer"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "groupID": {
                    "type": "integer"
                },
                "groupName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "integrationCode": {
                    "type": "string"
                },
                "isPOSDefaultCustomer": {
                    "type": "integer"
                },
                "lastModified": {
                    "type": "integer"
                },
                "lastModifierUsername": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "mailEnabled": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "operatorIdentifier": {
                    "type": "string"
                },
                "payerID": {
                    "type": "integer"
                },
                "paymentDays": {
                    "type": "integer"
                },
                "paysViaFactoring": {
                    "type": "integer"
                },
                "personTitleID": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "posCouponsDisabled": {
                    "type": "integer"
                },
                "postalCode": {
                    "type": "string"
                },
                "priceListID": {
                    "description": "Detailed info",
                    "type": "integer"
                },
                "priceListID2": {
                    "type": "integer"
                },
                "priceListID3": {
                    "type": "integer"
                },
                "referenceNumber": {
                    "type": "string"
                },
                "rewardPointsDisabled": {
                    "type": "integer"
                },
                "salesBlocked": {
                    "type": "integer"
                },
                "shipGoodsWithWaybills": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "taxExempt": {
                    "type": "integer"
                },
                "twitterID": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                },
                "vatNumber": {
                    "type": "string"
                },
                "webshopLastLogin": {
                    "type": "string"
                },
                "webshopUsername": {
                    "description": "Web-shop related fields",
                    "type": "string"
                }
            }
        },
        "models.UpdateCustomerRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id: binding:": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        }
    }
}